integerize the dataloader_batch_sz

+++ code/scripts/cluster/cluster_norm.py
@@ -123,7 +123,7 @@ else:
 
 config.out_dir = os.path.join(config.out_root, str(config.model_ind))
 assert (config.batch_sz % config.num_dataloaders == 0)
-config.dataloader_batch_sz = config.batch_sz / config.num_dataloaders
+config.dataloader_batch_sz = int(config.batch_sz / config.num_dataloaders)
 
 assert (config.mode == "IID")
 assert ("TwoHead" in config.arch)
